# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mission API",
    "version": "1.0"
  },
  "paths": {
    "/missions": {
      "get": {
        "summary": "List all missions",
        "responses": {
          "200": { "description": "A list of missions" }
        }
      },
      "post": {
        "summary": "Create a new mission",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "missionName": { "type": "string" },
                  "missionDescription": { "type": "string" }
                },
                "required": ["missionName"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Mission created" },
          "400": { "description": "Missing or invalid missionName" }
        }
      }
    },
    "/missions/{id}": {
      "get": {
        "summary": "Get a single mission by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Mission found" },
          "404": { "description": "Mission not found" }
        }
      },
      "put": {
        "summary": "Update a mission",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "missionName": { "type": "string" },
                  "missionDescription": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Mission updated" },
          "400": { "description": "Empty or invalid body" },
          "404": { "description": "Mission not found" }
        }
      },
      "delete": {
        "summary": "Delete a mission",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "Mission deleted" },
          "404": { "description": "Mission not found" }
        }
      }
    }
  }
}